Revert everything and start anew
