+++ tests/check_consistent.py
@@ -9,12 +9,13 @@ import filecmp
 consistent_files = [
     {'stdlib/2/builtins.pyi', 'stdlib/2/__builtin__.pyi'},
     {'stdlib/2/SocketServer.pyi', 'stdlib/3/socketserver.pyi'},
-    {'stdlib/2/os2emxpath.pyi', 'stdlib/2/posixpath.pyi', 'stdlib/2/ntpath.pyi', 'stdlib/2/macpath.pyi', 'stdlib/2/os/path.pyi'},
+    {'stdlib/2/os2emxpath.pyi', 'stdlib/2/posixpath.pyi', 'stdlib/2/ntpath.pyi', 'stdlib/2/macpath.pyi'},
     {'stdlib/2and3/pyexpat/__init__.pyi', 'stdlib/2and3/xml/parsers/expat/__init__.pyi'},
     {'stdlib/2and3/pyexpat/errors.pyi', 'stdlib/2and3/xml/parsers/expat/errors.pyi'},
     {'stdlib/2and3/pyexpat/model.pyi', 'stdlib/2and3/xml/parsers/expat/model.pyi'},
-    {'stdlib/3/ntpath.pyi', 'stdlib/3/posixpath.pyi', 'stdlib/3/macpath.pyi', 'stdlib/3/posixpath.pyi', 'stdlib/3/os/path.pyi'},
+    {'stdlib/3/ntpath.pyi', 'stdlib/3/posixpath.pyi', 'stdlib/3/macpath.pyi', 'stdlib/3/posixpath.pyi'},
     {'stdlib/3/enum.pyi', 'third_party/2/enum.pyi'},
+    {'stdlib/2/os/path.pyi', 'stdlib/3/os/path.pyi'},
     {'stdlib/3/unittest/mock.pyi', 'third_party/2and3/mock.pyi'},
     {'stdlib/3/concurrent/__init__.pyi', 'third_party/2/concurrent/__init__.pyi'},
     {'stdlib/3/concurrent/futures/__init__.pyi', 'third_party/2/concurrent/futures/__init__.pyi'},
